+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/verification/v2/SelfVerificationSteps.kt
@@ -68,7 +68,7 @@ fun SelfVerificationSteps(viewModel: SelfVerificationViewModel) {
         if (availableSelfVerificationMethods == null) {
             WizardSection {
                 LoadingSpinner()
-                Text(i18n.verificationWait())
+                Text(i18n.selfVerificationWaitingForMethods())
             }
         } else {
             VerifyWithOtherDevice(viewModel)